Updated the prompt bar; adding prototype animations and sound effects.
Changed the prompt bar to be words instead of pictures.
Altered and added scripts to play an audio source.
The scripts were added to the game objects and their word to prepare us for when we have official audio.
A green check mark is placed over their word when the game object is clicked, instead of destroying the word.
Clearing the level of game objects returns the user to the main menu rather than closing the application.
A prototype animation was added of a ball of yarn rolling across the screen.
I wasn't able to find a good image for the yarn.
The yarn animation has a function to return to its original position.
This function can be referenced after the hint meter has filled, and Lilly has given a hint.
For now, it just resets when it reaches its target position.
The yarns movement can be altered to match the speed of the hint bar.