Removed glow on "Cats" button. Cats are blacked out, rather than invisible.
I couldn't get an effect that I did like, so I scrapped the idea for flashing the "cats" button right now.
I thought it would be better for the players to have an idea of what the new cats are, so I made them blackened instead of invisible until the player unlocks them.