"RateMe" popup
Doing the rating prompt was as simple as Mark's link to the barrie patch website. Application.OpenURL("whatever site")

After the player pushes the yes button, PlayerPrefs("haveRated") is set to 1 and the prompt only opens if "haveRated"==0.

Prompt also only opens after level two is completed, this is a bit of time for the player to get acquainted. It can be changed.

Prompt always opens if level two is completed and player has not yet rated, idk if this is too annoying or in their face. Maybe if it opened on a later level, or if I put in a delay until it opens again?