Updated art assests, updated credits popup, added delay to "rateMe" popup.
Added new art assets to the hard levels, and rearranged what I could into some potential placement drafts.
Added more detail to the credits popup.
Added a countdown to the popup.

Within RateMe.cs:
PlayerPrefs of "rateDelay" is checked to be <=0. This means the countdown is ready and the script can continue.

On Start(), delayCountdown is set to "rateDelay".

delayCountdown is minused by one.

"rateDelay" is set to the new delayCountdown creating a countdown for the script.

When the user closes the popup, "rateDelay" is set to 4, so that the popup opens only every 4 times the scene is loaded.